requirements.txt
Pillow
# Add runtime deps here (e.g., RPi.GPIO) when deploying to Raspberry Pi
